Fix crash when opening the Game Launcher option window
This happens when the elevated user ( usually Administrator ) has not set the MIDI device because it never run FF7Config as admin before.
The code has already a defensive code practice but the if condition was set wrong. By fixing that we default on the first one if one is not currently set.