RuntimeVar: Add variable RandomVar that holds a random value
Other than adding a new variable RandomVar, the RandomVarOnce has been updated to use the same data structure of RandomVar.
The idea behind this variable is to have a variable that holds a random value that can be rerolled. The reroll strategy is based on two conditions:
 - The lookup of the conditional folder of this `RuntimeVar` has just started (this is controlled by the `locked` attribute, which is locked after the first lookup and unlocked after it is resolved)
 - A delta time in seconds has been passed from the last reroll
If these two conditions are satisfied, the random variable is rerolled, and the last reroll timestamp is updated.
This can be used with the notation:
```xml
<RuntimeVar Var="RandomVar:{varName}:{size}:{deltaTime}" Values="217..218" />
```
where:
 - `{varName}` has to be substituted with an identifier name for the variable
 - `{size}` has to be substituted with a number that is the range of the random variable that goes from [0, size-1]
 - `{deltaTime}` is optional and has to be substituted with the minimum delta time to allow the random variable to be rerolled (if it's not set, `deltaTime` is 0)

Notice that you can combine `RandomVar` and `RandomVarOnce` to use the same identifier `{varName}`. In this way, when it looks up for `RandomVarOnce` it does not reroll anything, but only get the
random value.
Another two important prerequisites are:
 - Each conditional folder using the same `RandomVar` should contain the same files
 - All the values of the `RandomVar` should be mapped to a conditional folder (e.g. if you have `RandomVar:prelude:2` and there is only a conditional folder that looks for `Values=0`, this won't work)