SeventhHeavenUI: Auto manage use_external_music
This commit will remove the advanced flags for music and FFMpeg do reduce the confusion on the user.

The FFMpeg movie flag was unnecessary as for FF7 there was never the case to disable that in the first place.

The Music flag will instead be managed by looking if the current installation has ogg files in "music/vgmstream".
If not, it will fallback on using MIDI by default.
If a mod is installed, it will automatically override this flag using the mod.xml layer.