MyModsViewModel: Fix massive disable when installing a new mod
If the user uses the filter to search for a mod on the catalog view, this filter will trip the mod refresh panel, since the mod list will be filtered as well.
Disabling the filtering logic will allow to keep the entire mod list and add on top the new one in the right place, without disabling all the others.